refactor : useGiftRanking.tsx 파일을 components 폴더에서 hooks 폴더로 이동
